+++ core/client/hdfs/src/main/java/alluxio/hadoop/LocalCacheFileSystem.java
@@ -17,11 +17,11 @@ import static java.nio.charset.StandardCharsets.UTF_8;
 import alluxio.AlluxioURI;
 import alluxio.CloseableSupplier;
 import alluxio.client.file.CacheContext;
+import alluxio.client.file.PositionReadFileInStream;
 import alluxio.client.file.URIStatus;
 import alluxio.client.file.cache.CacheManager;
 import alluxio.client.file.cache.LocalCachePositionReader;
 import alluxio.client.file.cache.filter.CacheFilter;
-import alluxio.client.file.dora.PositionReadFileInStream;
 import alluxio.conf.AlluxioConfiguration;
 import alluxio.conf.PropertyKey;
 import alluxio.exception.runtime.AlluxioRuntimeException;
